Reorganize 'mr-jar' application tests
