+++ appserver/tests/application/src/main/java/org/glassfish/main/test/app/mrjar/VersionImpl.java
@@ -14,7 +14,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package org.glassfish.main.test.app.web.mrjar;
+package org.glassfish.main.test.app.mrjar;
 
 /**
  * This is a stub implementation used only for creation of the test application.
