+++ appserver/tests/application/src/main/java/org/glassfish/main/test/app/mrjar/servlet/MultiReleaseServlet.java
@@ -14,7 +14,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package org.glassfish.main.test.app.web.mrjar;
+package org.glassfish.main.test.app.mrjar.servlet;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
@@ -23,6 +23,8 @@ import jakarta.servlet.http.HttpServletResponse;
 
 import java.io.IOException;
 import java.io.PrintWriter;
+
+import org.glassfish.main.test.app.mrjar.Version;
 
 public class MultiReleaseServlet extends HttpServlet {
