+++ appserver/tests/application/src/main/java/org/glassfish/main/test/app/mrjar/servlet/MultiReleaseServletContainerInitializer.java
@@ -14,7 +14,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package org.glassfish.main.test.app.web.mrjar;
+package org.glassfish.main.test.app.mrjar.servlet;
 
 import jakarta.servlet.ServletContainerInitializer;
 import jakarta.servlet.ServletContext;
@@ -23,6 +23,8 @@ import jakarta.servlet.ServletRegistration;
 import jakarta.servlet.annotation.HandlesTypes;
 
 import java.util.Set;
+
+import org.glassfish.main.test.app.mrjar.Version;
 
 @HandlesTypes({Version.class})
 public class MultiReleaseServletContainerInitializer implements ServletContainerInitializer {
