+++ appserver/tests/application/src/test/java/org/glassfish/main/test/app/mrjar/ear/MultiReleaseEnterpriseApplicationLibraryTest.java
@@ -14,7 +14,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package org.glassfish.main.test.app.web.mrjar;
+package org.glassfish.main.test.app.mrjar.ear;
 
 import java.io.File;
 import java.io.IOException;
@@ -22,8 +22,11 @@ import java.net.HttpURLConnection;
 import java.nio.file.Files;
 
 import org.glassfish.main.itest.tools.GlassFishTestEnvironment;
+import org.glassfish.main.test.app.mrjar.MultiReleaseTestBase;
+import org.glassfish.main.test.app.mrjar.webapp.MultiReleaseApplication;
+import org.glassfish.main.test.app.mrjar.webapp.MultiReleaseResource;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
-import org.jboss.shrinkwrap.api.exporter.ZipExporter;
+import org.jboss.shrinkwrap.api.spec.EnterpriseArchive;
 import org.jboss.shrinkwrap.api.spec.JavaArchive;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.junit.jupiter.api.AfterAll;
@@ -44,39 +47,40 @@ import static org.junit.jupiter.api.condition.JRE.JAVA_17;
 import static org.objectweb.asm.Opcodes.V11;
 import static org.objectweb.asm.Opcodes.V17;
 
-public class MultiReleaseWebApplicationTest extends MultiReleaseTestBase{
+public class MultiReleaseEnterpriseApplicationLibraryTest extends MultiReleaseTestBase {
 
-    private static final System.Logger LOG = System.getLogger(MultiReleaseWebApplicationTest.class.getName());
+    private static final System.Logger LOG = System.getLogger(MultiReleaseEnterpriseApplicationLibraryTest.class.getName());
 
-    private static final String APP_NAME = "mrwebapp";
+    private static final String APP_NAME = "mrear";
 
-    private static final String CONTEXT_ROOT = "/" + APP_NAME;
+    private static final String EAR_FILE_NAME = APP_NAME + ".ear";
 
-    private static final String APP_FILE_NAME = APP_NAME + ".war";
+    private static final String WAR_FILE_NAME = "mrwar.war";
+
+    private static final String CONTEXT_ROOT = "/mrwar";
 
     @BeforeAll
     public static void deploy() throws IOException {
-        File warFile = createDeployment();
+        File earFile = createDeployment();
         try {
-            assertThat(ASADMIN.exec("deploy", warFile.getAbsolutePath()), asadminOK());
+            assertThat(ASADMIN.exec("deploy", earFile.getAbsolutePath()), asadminOK());
         } finally {
             try {
-                Files.deleteIfExists(warFile.toPath());
+                Files.deleteIfExists(earFile.toPath());
             } catch (IOException e) {
-                LOG.log(WARNING, "An error occurred while remove temporary file " + warFile.getAbsolutePath(), e);
+                LOG.log(WARNING, "An error occurred while remove temporary file " + earFile.getAbsolutePath(), e);
             }
-
         }
     }
 
     @AfterAll
     public static void undeploy() {
-        assertThat(ASADMIN.exec("undeploy", APP_NAME), asadminOK());
+        assertThat(ASADMIN.exec("undeploy", "mrear"), asadminOK());
     }
 
     @Test
     @EnabledForJreRange(min = JAVA_11, max = JAVA_16)
-    public void testMultiReleaseJarProcessingJdk11(TestInfo testInfo) throws IOException {
+    public void testMultiReleaseEarLibProcessingJdk11(TestInfo testInfo) throws IOException {
         LOG.log(INFO, "Run test method {0}", testInfo.getTestMethod().orElseThrow().getName());
         HttpURLConnection connection = GlassFishTestEnvironment.openConnection(8080, CONTEXT_ROOT);
         connection.setRequestMethod("GET");
@@ -93,7 +97,7 @@ public class MultiReleaseWebApplicationTest extends MultiReleaseTestBase{
 
     @Test
     @EnabledForJreRange(min = JAVA_17)
-    public void testMultiReleaseJarProcessingJdk17(TestInfo testInfo) throws IOException {
+    public void testMultiReleaseEarLibProcessingJdk17(TestInfo testInfo) throws IOException {
         LOG.log(INFO, "Run test method {0}", testInfo.getTestMethod().orElseThrow().getName());
         HttpURLConnection connection = GlassFishTestEnvironment.openConnection(8080, CONTEXT_ROOT);
         connection.setRequestMethod("GET");
@@ -109,19 +113,18 @@ public class MultiReleaseWebApplicationTest extends MultiReleaseTestBase{
     }
 
     private static File createDeployment() throws IOException {
-        // Create Multi-Release JAR library
-        JavaArchive jarArchive = createMultiReleaseLibrary();
+        JavaArchive javaArchive = createMultiReleaseLibrary();
 
-        // Create test web application archive
-        WebArchive webArchive = ShrinkWrap.create(WebArchive.class)
-            .addAsLibrary(jarArchive)
+        WebArchive webArchive = ShrinkWrap.create(WebArchive.class, WAR_FILE_NAME)
             .addClass(MultiReleaseResource.class)
             .addClass(MultiReleaseApplication.class);
 
-        File tmpDir = Files.createTempDirectory(APP_NAME).toFile();
-        File warFile = new File(tmpDir, APP_FILE_NAME);
-        webArchive.as(ZipExporter.class).exportTo(warFile, true);
-        tmpDir.deleteOnExit();
-        return warFile;
+        EnterpriseArchive enterpriseArchive = ShrinkWrap.create(EnterpriseArchive.class, EAR_FILE_NAME)
+            .addAsLibrary(javaArchive)
+            .addAsModule(webArchive);
+
+        LOG.log(INFO, enterpriseArchive.toString(true));
+
+        return createFileFor(enterpriseArchive, APP_NAME);
     }
 }
