+++ appserver/tests/application/src/test/java/org/glassfish/main/test/app/mrjar/webapp/MultiReleaseWebApplicationTest.java
@@ -14,7 +14,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package org.glassfish.main.test.app.web.mrjar;
+package org.glassfish.main.test.app.mrjar.webapp;
 
 import java.io.File;
 import java.io.IOException;
@@ -22,9 +22,8 @@ import java.net.HttpURLConnection;
 import java.nio.file.Files;
 
 import org.glassfish.main.itest.tools.GlassFishTestEnvironment;
+import org.glassfish.main.test.app.mrjar.MultiReleaseTestBase;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
-import org.jboss.shrinkwrap.api.exporter.ZipExporter;
-import org.jboss.shrinkwrap.api.spec.EnterpriseArchive;
 import org.jboss.shrinkwrap.api.spec.JavaArchive;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.junit.jupiter.api.AfterAll;
@@ -45,40 +44,36 @@ import static org.junit.jupiter.api.condition.JRE.JAVA_17;
 import static org.objectweb.asm.Opcodes.V11;
 import static org.objectweb.asm.Opcodes.V17;
 
-public class MultiReleaseEnterpriseApplicationLibraryTest extends MultiReleaseTestBase {
+public class MultiReleaseWebApplicationTest extends MultiReleaseTestBase {
 
-    private static final System.Logger LOG = System.getLogger(MultiReleaseEnterpriseApplicationLibraryTest.class.getName());
+    private static final System.Logger LOG = System.getLogger(MultiReleaseWebApplicationTest.class.getName());
 
-    private static final String APP_NAME = "mrear";
+    private static final String APP_NAME = "mrwebapp";
 
-    private static final String EAR_FILE_NAME = APP_NAME + ".ear";
-
-    private static final String WAR_FILE_NAME = "mrwar.war";
-
-    private static final String CONTEXT_ROOT = "/mrwar";
+    private static final String CONTEXT_ROOT = "/" + APP_NAME;
 
     @BeforeAll
     public static void deploy() throws IOException {
-        File earFile = createDeployment();
+        File warFile = createDeployment();
         try {
-            assertThat(ASADMIN.exec("deploy", earFile.getAbsolutePath()), asadminOK());
+            assertThat(ASADMIN.exec("deploy", warFile.getAbsolutePath()), asadminOK());
         } finally {
             try {
-                Files.deleteIfExists(earFile.toPath());
+                Files.deleteIfExists(warFile.toPath());
             } catch (IOException e) {
-                LOG.log(WARNING, "An error occurred while remove temporary file " + earFile.getAbsolutePath(), e);
+                LOG.log(WARNING, "An error occurred while remove temporary file " + warFile.getAbsolutePath(), e);
             }
         }
     }
 
     @AfterAll
     public static void undeploy() {
-        assertThat(ASADMIN.exec("undeploy", "mrear"), asadminOK());
+        assertThat(ASADMIN.exec("undeploy", APP_NAME), asadminOK());
     }
 
     @Test
     @EnabledForJreRange(min = JAVA_11, max = JAVA_16)
-    public void testMultiReleaseEarLibProcessingJdk11(TestInfo testInfo) throws IOException {
+    public void testMultiReleaseJarProcessingJdk11(TestInfo testInfo) throws IOException {
         LOG.log(INFO, "Run test method {0}", testInfo.getTestMethod().orElseThrow().getName());
         HttpURLConnection connection = GlassFishTestEnvironment.openConnection(8080, CONTEXT_ROOT);
         connection.setRequestMethod("GET");
@@ -95,13 +90,14 @@ public class MultiReleaseEnterpriseApplicationLibraryTest extends MultiReleaseTe
 
     @Test
     @EnabledForJreRange(min = JAVA_17)
-    public void testMultiReleaseEarLibProcessingJdk17(TestInfo testInfo) throws IOException {
+    public void testMultiReleaseJarProcessingJdk17(TestInfo testInfo) throws IOException {
         LOG.log(INFO, "Run test method {0}", testInfo.getTestMethod().orElseThrow().getName());
         HttpURLConnection connection = GlassFishTestEnvironment.openConnection(8080, CONTEXT_ROOT);
         connection.setRequestMethod("GET");
         try {
             assertAll(
                 () -> assertThat(connection.getResponseCode(), equalTo(200)),
+                // Check version of loaded class file
                 () -> assertThat(Integer.parseInt(readResponse(connection)), equalTo(V17))
             );
         } finally {
@@ -110,22 +106,17 @@ public class MultiReleaseEnterpriseApplicationLibraryTest extends MultiReleaseTe
     }
 
     private static File createDeployment() throws IOException {
-        JavaArchive javaArchive = createMultiReleaseLibrary();
+        // Create Multi-Release JAR library
+        JavaArchive jarArchive = createMultiReleaseLibrary();
 
-        WebArchive webArchive = ShrinkWrap.create(WebArchive.class, WAR_FILE_NAME)
+        // Create test web application archive
+        WebArchive webArchive = ShrinkWrap.create(WebArchive.class)
+            .addAsLibrary(jarArchive)
             .addClass(MultiReleaseResource.class)
             .addClass(MultiReleaseApplication.class);
 
-        EnterpriseArchive enterpriseArchive = ShrinkWrap.create(EnterpriseArchive.class, EAR_FILE_NAME)
-            .addAsLibrary(javaArchive)
-            .addAsModule(webArchive);
+        LOG.log(INFO, webArchive.toString(true));
 
-        LOG.log(INFO, enterpriseArchive.toString(true));
-
-        File tmpDir = Files.createTempDirectory(APP_NAME).toFile();
-        File earFile = new File(tmpDir, EAR_FILE_NAME);
-        enterpriseArchive.as(ZipExporter.class).exportTo(earFile, true);
-        tmpDir.deleteOnExit();
-        return earFile;
+        return createFileFor(webArchive, APP_NAME);
     }
 }
